Project Merge/requirements.txt
fastapi
uvicorn[standard]
pymysql
python-multipart
jinja2
passlib[bcrypt]
python-dotenv
sentence-transformers
transformers
scikit-learn
numpy
pymupdf
pillow
pytesseract
pythainlp
